% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_database.R
\name{connect_database}
\alias{connect_database}
\title{Query Bioinformatics Database}
\usage{
connect_database(species_id = NULL)
}
\arguments{
\item{species_id}{ID of species selected (Loads organism info data if NULL)}
}
\value{
SQLite connection to the downloaded file
}
\description{
Retrieves database (.db file) from the SDSU bioinformatics database,
 creates a connection via SQLite
}
