% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finch_sample}
\alias{finch_sample}
\title{Sample of Zebra Finch Gene Counts}
\format{
\subsection{\code{finch_sample}}{

A data frame with 100 rows and 19 columns
\describe{
\item{Row Names}{Gene Name/Label}
\item{Columns}{Gene count readings for sepearate groups}
\item{Experimental Groups}{
\itemize{
\item Male vs. Female
\item Heat vs. Control}
}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
finch_sample
}
\description{
A subset of Zebra Finch (Taeniopygia guttata) gene count readings of 21,407
genes. The subset contains 100 randomly sampled genes from 19 individuals
from different experimental groups.
}
\keyword{datasets}
