% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_filter.R
\name{gene_filter}
\alias{gene_filter}
\title{Filter Gene Data}
\usage{
gene_filter(data, chrs = NULL, min_gc_content = NULL, max_gc_content = NULL)
}
\arguments{
\item{data}{Data frame containing gene IDs and information}

\item{chrs}{Vector of chromosome names/numbers to keep in data}

\item{min_gc_content}{Minimum percentage GC content for genes}

\item{max_gc_content}{Maximum percentage GC content for genes}
}
\value{
Data frame of filtered gene information
}
\description{
Filter Gene Data
}
