% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get Table from Selected Database}
\usage{
get_table(species_id = NULL, srch_species = FALSE, table = NULL)
}
\arguments{
\item{species_id}{Selected species ID (Loads organism info data if NULL)}

\item{srch_species}{TRUE/FALSE if function is being called from srch_species
function, FALSE by default}

\item{table}{Selected table, NULL by default}
}
\value{
Returns data frame selected from table list
}
\description{
Prints list of tables for selected species and allows user to select
 from the list, returning the selected table
}
