% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_T2G}
\alias{hypox_T2G}
\title{Example KEGG Pathway Mapping for Hypoxia}
\format{
A data frame with 2 columns:
\describe{
  \item{description}{Pathway ID or description}
  \item{gene}{Ensembl Gene ID}
}
}
\usage{
hypox_T2G
}
\description{
An example TERM2GENE mapping tailored for the `hypox_deseq` dataset,
demonstrating the `T2G_prep` function's output.
}
\keyword{datasets}
