% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypoxia_reads}
\alias{hypoxia_reads}
\title{Hypoxia Data Gene Counts}
\format{
\subsection{\code{hypoxia_reads}}{

A data frame with 35,238 rows and 4 columns:
\describe{
\item{MRS1220_hypoxia_rep1}{Treatment replication 1 counts}
\item{MRS1220_hypoxia_rep2}{Treatment replication 2 counts}
\item{vehicle_hypoxia_rep1}{Control replication 1 counts}
\item{vehicle_hypoxia_rep2}{Control replication 2 counts}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE100146}
}
\usage{
hypoxia_reads
}
\description{
Gene counts gathered in the following experiment: RNAseq transcriptomic profile of
glioblastoma stem-like cells derived from U87MG cell line treated with a
selective A3 adenosine receptor antagonist (MRS1220) under hypoxia.
}
\keyword{datasets}
