% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_cat.R
\name{path_cat}
\alias{path_cat}
\title{Retrieve Pathway Categories}
\usage{
path_cat(species_id = NULL)
}
\arguments{
\item{species_id}{Numeric. The ID of a desired species from database, found
using `srch_species()`}
}
\value{
Data frame of pathway categories for given species
}
\description{
Retrieves pathway category options (e.g. KEGG, GOBP, etc.) for a given
species. May take longer for well-documented species (i.e. Human)
}
