% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Gene Expression Data}
\usage{
process_data(
  data,
  missing_value = "geneMedian",
  min_cpm = 0.5,
  n_min_samples_count = 1,
  counts_transform = 0,
  counts_log_start = 4
)
}
\arguments{
\item{data}{A numeric matrix or data frame of gene expression counts.}

\item{missing_value}{Character. Method to handle missing values. Options:
\itemize{
\item \code{"geneMedian"}: Impute using the median expression of the gene across
samples.
\item \code{"treatAsZero"}: Replace NAs with 0.
\item \code{"groupMedian"}: Impute using the median of the sample group
(detected from colnames).
}}

\item{min_cpm}{Numeric. Minimum counts per million threshold for filtering
genes.}

\item{n_min_samples_count}{Numeric. Minimum number of samples that must meet
the \code{min_cpm} threshold for a gene to be retained.}

\item{counts_transform}{Integer. Method for data transformation:
0. None (Raw Counts).
\enumerate{
\item log2(CPM + \code{counts_log_start})
\item Variance Stabilizing Transformation (VST) via \code{DESeq2}.
\item Regularized Log (rlog) via \code{DESeq2}.
}}

\item{counts_log_start}{Numeric. Constant added to counts before log
transformation. Usually between 1 and 10. Higher values reduce noise but
decrease sensitivity
(used when \code{counts_transform = 1}).}
}
\value{
The processed and transformed data matrix.
}
\description{
Performs pre-processing, missing value imputation, filtering, and
transformation on gene expression count data.
}
