% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_id.R
\name{convert_id}
\alias{convert_id}
\title{Convert gene IDs to ensembl data.}
\usage{
convert_id(genes, data = NULL, species_id, max_sample_ids = 100)
}
\arguments{
\item{data}{}

\item{max_sample_ids}{The number of gene ids used to determine species}
}
\value{
A large list of the conversion ID information that was gathered
 from querying the database with the original IDs.
}
\description{
Convert gene IDs to ensembl data.
}
