% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathways.R
\name{get_pathways}
\alias{get_pathways}
\title{Get Species Pathways}
\usage{
get_pathways(species_id, genes = NULL)
}
\arguments{
\item{species_id}{ID of desired species}

\item{genes}{Vector or column of data frame containing genes of interest}
}
\value{
Data frame of pathways that contain the provided genes (returns
 all pathways if genes is NULL)
}
\description{
Get Species Pathways
}
