% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List Table Options}
\usage{
list_tables(species_id = NULL)
}
\arguments{
\item{species_id}{Selected species ID (Loads organism info data if NULL)}
}
\value{
Returns a list of table names
}
\description{
List Table Options
}
