% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Gene Expression Data}
\usage{
process_data(
  data,
  missing_value = "geneMedian",
  min_cpm = 0.5,
  n_min_samples_count = 1,
  counts_transform = 0,
  counts_log_start = 4
)
}
\arguments{
\item{data}{A numeric matrix or data frame of gene expression counts.}

\item{missing_value}{Character. Method to handle missing values. Options:
\itemize{
\item \code{"geneMedian"}: Impute using the median expression of the gene across
samples.
\item \code{"treatAsZero"}: Replace NAs with 0.
\item \code{"groupMedian"}: Impute using the median of the sample group
(detected from colnames).
}}

\item{min_cpm}{Numeric. Minimum counts per million threshold for filtering
genes.}

\item{n_min_samples_count}{Numeric. Minimum number of samples that must meet
the \code{min_cpm} threshold for a gene to be retained.}

\item{counts_transform}{Integer. Method for data transformation:
\itemize{
\item 0 = None (Raw Counts).
\item 1 = log2(CPM + \code{counts_log_start})
\item 2 = Variance Stabilizing Transformation (VST) via \code{DESeq2}.
\item 3 = Regularized Log (rlog) via \code{DESeq2}.
}}

\item{counts_log_start}{Numeric. Constant added to counts before log
transformation. Usually between 1 and 10. Higher values reduce noise but
decrease sensitivity
(used when \code{counts_transform = 1}).}
}
\value{
The processed and transformed data matrix.
}
\description{
Performs pre-processing, missing value imputation, filtering, and
transformation on gene expression count data.
}
\examples{
# Check example data
summary(pathdb::hypoxia_reads)
nrow(pathdb::hypoxia_reads)

# YOU decide how your data is transformed.
# Here, we want to:
# Replace missing values with median
# Set minimum counts-per-million of 0.4
# Meet CPM threshold in 2 samples
# Keep raw counts

hypox_filtered <- process_data(data = pathdb::hypoxia_reads,
                               missing_value = "geneMedian",
                               min_cpm = 0.4,
                               n_min_samples_count = 2,
                               counts_transform = 0)

# Check filtered data
summary(hypox_filtered)
nrow(hypox_filtered)

}
