% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srch_species.R
\name{srch_species}
\alias{srch_species}
\title{Search for Species by Name}
\usage{
srch_species(query, name_type)
}
\arguments{
\item{query}{Species name or partial name}

\item{name_type}{Name type for species ("common", "academic", "id); Be aware
that common names are not available for all species}
}
\value{
Data frame of all species info with names that match the query
}
\description{
Search for Species by Name
}
